se agrega funcionalidad para preguntar si es necesario eliminar una lista de tareas

+++ organizador-tareas/src/app/board/board/board.component.ts
@@ -6,6 +6,7 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { BoardDetalle } from 'src/app/clases/boardDetalle.interface';
 import { NgxSpinnerService } from 'ngx-spinner';
 import { ColumnasService } from 'src/app/Servicios/columnas.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-board',
@@ -17,7 +18,7 @@ export class BoardComponent implements OnInit {
   tablero: any = [];
   nombreProyecto: String;
   public codigoTablero: string | null;
-  
+
 
   constructor(
     public boardService: BoardService,
@@ -74,7 +75,24 @@ export class BoardComponent implements OnInit {
   }
 
   onDeleteColumn(columnId: number) {
-    this.boardService.deleteColumn(columnId, this.codigoTablero)
+    Swal.fire({
+      title: '¿Está seguro que desea eliminar esta lista con todas sus tareas?',
+      showDenyButton: true,
+      showConfirmButton: false,
+      showCancelButton: true,
+      denyButtonText: `Eliminar`,
+    }).then((result) => {
+      /* Read more about isConfirmed, isDenied below */
+      if (result.isDenied) {
+        this.boardService.deleteColumn(columnId, this.codigoTablero);
+        Swal.fire({
+          title: 'Lista de Tareas eliminada con éxito',
+          icon: 'success'
+        }
+        )
+      }
+    })
+    
   }
 
   onDeleteCard(cardId: number, columnId: number) {
